Add the --preserve-case option.
While Active Directory is case-aware but not case-sensitive, passwd(5)
and group(5) files are often case-sensitive, and downstream consumers
of our generated passwd/group may treat them as case-sensitive.
(E.g., the username "SOMEHOST$" may not match "somehost$".)

So while we still follow sssd, and default to outputting our generated
passwd/group files in all lowercase, the --preserve-case option can be
used to preserve case for downstream consumers that are
case-sensitive.